solved part 6 of set 1

+++ src/set1/ex6_tests.rs
@@ -5,7 +5,7 @@ mod transpose_blocks_tests {
     #[test]
     fn test_transpose_blocks_empty_input() {
         let blocks: Vec<Vec<u8>> = vec![];
-        let keylength: u32 = 4;
+        let keylength: usize = 4;
         let result = transpose_blocks(&blocks, keylength);
 
         let expected: Vec<Vec<u8>> = vec![vec![]; keylength as usize];
@@ -15,7 +15,7 @@ mod transpose_blocks_tests {
     #[test]
     fn test_transpose_blocks_single_block() {
         let blocks: Vec<Vec<u8>> = vec![vec![1, 2, 3, 4]];
-        let keylength: u32 = 4;
+        let keylength: usize = 4;
         let result = transpose_blocks(&blocks, keylength);
 
         let expected: Vec<Vec<u8>> = vec![vec![1], vec![2], vec![3], vec![4]];
@@ -25,7 +25,7 @@ mod transpose_blocks_tests {
     #[test]
     fn test_transpose_blocks_multiple_blocks() {
         let blocks: Vec<Vec<u8>> = vec![vec![1, 2, 3, 4], vec![5, 6, 7, 8], vec![9, 10, 11, 12]];
-        let keylength: u32 = 4;
+        let keylength: usize = 4;
         let result = transpose_blocks(&blocks, keylength);
 
         let expected: Vec<Vec<u8>> = vec![
@@ -40,7 +40,7 @@ mod transpose_blocks_tests {
     #[test]
     fn test_transpose_blocks_uneven_blocks() {
         let blocks: Vec<Vec<u8>> = vec![vec![1, 2, 3, 4], vec![5, 6, 7], vec![9, 10, 11, 12]];
-        let keylength: u32 = 4;
+        let keylength: usize = 4;
         let result = transpose_blocks(&blocks, keylength);
 
         let expected: Vec<Vec<u8>> =
@@ -51,7 +51,7 @@ mod transpose_blocks_tests {
     #[test]
     fn test_transpose_blocks_varying_keylength() {
         let blocks: Vec<Vec<u8>> = vec![vec![1, 2, 3, 4], vec![5, 6, 7, 8], vec![9, 10, 11, 12]];
-        let keylength: u32 = 2;
+        let keylength: usize = 2;
         let result = transpose_blocks(&blocks, keylength);
 
         let expected: Vec<Vec<u8>> = vec![vec![1, 5, 9], vec![2, 6, 10]];
